requirements.txt
appnope==0.1.3
asttokens==2.4.1
azure-cognitiveservices-speech==1.33.0
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
importlib-metadata==6.9.0
ipykernel==6.27.1
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jupyter-client==8.6.0
jupyter-core==5.5.0
matplotlib-inline==0.1.6
msal==1.25.0
msal-extensions==1.0.0
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
platformdirs==4.0.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
swagger-client==1.0.0
tornado==6.4
traitlets==5.14.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
